Update project.lisp
Added outputs for the status of the current Agent playing, including his bet, the current cards on the played so far by all agents (including the current player's hand), the current pot, and the communal cards.  
Need to consolidate outputs so that only the cards played for the given round are shown along with the communal cards and so that if an agent folds, there is a nil place-holder in the list of all cards played.